Re-organized react file structure
