--- skills_DS/react/src/components/App.js
@@ -1,8 +1,0 @@
-
-function App() {
-  return (
-   <h1>Hello World!</h1> 
-  );
-}
-
-export default App;
